Replace auto-wrap so metadata can be passed
Removes Page & Layout, replacing them with Wrapper with optional title, description props for metadata
