+++ src/components/home/Intro.tsx
@@ -6,7 +6,7 @@ import useMeasure, { UseMeasureRef } from 'react-use/lib/useMeasure'
 import LogoSvg from '../../images/mstable-logo.svg'
 import { Section } from '../layout/Section'
 import { MainAndAside } from '../layout/Grid'
-import { useLogoVisibility } from '../layout/Page'
+import { useLogoVisibility } from '../layout/Wrapper'
 import { FullBleed } from '../layout/FullBleed'
 import { BigStats } from './BigStats'
 import { AppCTA } from '../CTA'
