--- src/components/layout/Layout.tsx
@@ -1,11 +1,0 @@
-import React, { FC } from 'react'
-
-import '../../styles/index.styl'
-import { SEO } from './SEO'
-
-export const Layout: FC = ({ children }) => (
-  <>
-    <SEO />
-    {children}
-  </>
-)
