+++ src/components/layout/Navbar.tsx
@@ -5,7 +5,7 @@ import { Link } from 'gatsby'
 import M from '../../images/mstable-m.svg'
 import Ether from '../../images/ether-logo.svg'
 import { ExternalLink } from '../ExternalLink'
-import { useLogoVisibility } from './Page'
+import { useLogoVisibility } from './Wrapper'
 import { Constants } from '../../theme'
 
 const AppLink = styled(ExternalLink)`
