+++ src/pages/404.tsx
@@ -1,14 +1,13 @@
 import React, { FC } from 'react'
 
-import { Page } from '../components/layout/Page'
-
+import { Wrapper } from '../components/layout/Wrapper'
 import { Section } from '../components/layout/Section'
 
 const Home: FC = () => {
   return (
-    <Page>
+    <Wrapper>
       <Section title="404">Not found :-(</Section>
-    </Page>
+    </Wrapper>
   )
 }
