+++ src/posts/sample-post.mdx
@@ -1,6 +1,7 @@
 ---
 date: '2020-09-11'
 title: 'Sample title'
+description: 'Sample description here'
 author: 'author'
 type: post
 ---
