+++ src/strings.ts
@@ -1,0 +1,6 @@
+export const metadata = { 
+  title:
+    'mStable – Autonomous and non-custodial stablecoin infrastructure',
+  description:
+    'mStable unites stablecoins, lending and swapping into one standard.'
+}
